# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "效率仪表盘后端API，提供任务、提交、用户、仓库、项目、组织等数据管理接口",
        "title": "Efficiency Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9990",
    "basePath": "/api",
    "paths": {
...
        "/analysis/efficiency": {
            "get": {
                "description": "根据维度和ID查询提效分析结果，优先从PG缓存获取，无缓存时从ES实时计算",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efficiency"
                ],
                "summary": "获取提效分析结果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "维度(work_dir或repo)",
                        "name": "dimension",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "维度ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始日期(YYYYMMDD)",
                        "name": "startDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束日期(YYYYMMDD)",
                        "name": "endDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.EfficiencyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/analysis/efficiency/correct": {
            "put": {
                "description": "纠正AI估算天数等字段，记录纠正历史，并重新计算提效比例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efficiency"
                ],
                "summary": "纠正提效分析数据",
                "parameters": [
                    {
                        "description": "纠正请求",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CorrectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.EfficiencyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/analysis/efficiency/history": {
            "get": {
                "description": "查询指定维度和ID的纠正历史记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Efficiency"
                ],
                "summary": "获取纠正历史",
                "parameters": [
                    {
                        "type": "string",
                        "description": "维度(work_dir或repo)",
                        "name": "dimension",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "维度ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.EfficiencyHistoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.AIEstimatedInfo": {
            "type": "object",
            "properties": {
                "corrected_days": {
                    "type": "number"
                },
                "is_corrected": {
                    "type": "boolean"
                },
                "raw_days": {
                    "type": "number"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "main.ActualTimeInfo": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "total_code_lines": {
                    "type": "integer"
                },
                "total_lead_time_ms": {
                    "type": "integer"
                },
                "total_process_time_ms": {
                    "type": "integer"
                },
                "user_count": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.ActualTimeUser"
                    }
                }
            }
        },
        "main.ActualTimeUser": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "lead_time_ms": {
                    "type": "integer"
                },
                "process_time_ms": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "main.CorrectionHistoryItem": {
            "type": "object",
            "properties": {
                "corrected_at": {
                    "type": "string"
                },
                "corrected_by": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                },
                "new_value": {
                    "type": "string"
                },
                "old_value": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "main.CorrectionRequest": {
            "type": "object",
            "properties": {
                "by": {
                    "type": "string"
                },
                "dimension": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "main.CostInfo": {
            "type": "object",
            "properties": {
                "api_cost": {
                    "type": "number"
                },
                "cost_saving": {
                    "type": "number"
                },
                "daily_rate": {
                    "type": "number"
                },
                "roi": {
                    "type": "number"
                }
            }
        },
...
        "main.EfficiencyHistoryResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.CorrectionHistoryItem"
                    }
                }
            }
        },
        "main.EfficiencyInfo": {
            "type": "object",
            "properties": {
                "ratio_lead": {
                    "type": "number"
                },
                "ratio_process": {
                    "type": "number"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "main.EfficiencyResponse": {
            "type": "object",
            "properties": {
                "actual_time": {
                    "$ref": "#/definitions/main.ActualTimeInfo"
                },
                "ai_estimated": {
                    "$ref": "#/definitions/main.AIEstimatedInfo"
                },
                "analysis_date": {
                    "type": "string"
                },
                "analysis_file": {
                    "type": "string"
                },
                "cost": {
                    "$ref": "#/definitions/main.CostInfo"
                },
                "dimension": {
                    "type": "string"
                },
                "dimension_id": {
                    "type": "string"
                },
                "efficiency": {
                    "$ref": "#/definitions/main.EfficiencyInfo"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
    }
}